adding Day 3 assignments

+++ Day2/StudentProgram/Student.java
@@ -8,7 +8,7 @@ public class Student
 
 		int totalMarks = marks1+marks2+marks3;
 
-		int averageMarks = totalMarks/3;
+		float averageMarks = (float) totalMarks/3;
 
 		System.out.println("Total Marks of the student are : " + totalMarks);
 		System.out.println("Average Marks of the student are : " + averageMarks);
